Fix an orphaned prepared transaction case due to race between checkpointer and COMMIT PREPARE xlog recording
On Greenplum, checkpoint would collect prepared transactions which are actually
committed. If the COMMIT PREPARE xlog is before checkpoint.redo, after the
segment reboot, there would always be an orphaned (actually committed) prepared
transaction in memory. That happens when we collect the prepared transaction in
checkpointer before gxact->valid is reset and after the COMMIT PREPARE xlog is
recorded, see code in FinishPreparedTransaction().

That could lead to various issues. e.g. dtx recovery would keep trying to abort
that and then cause panic on the segment with message like "cannot abort
transaction 3285003, it was already committed (twophase.c:2205)".

Fixing this by adding a new variable committed in gxact to specifiy whether the
global transaction is committed or not. If being committed we surely do not
need to log the gxact in checkpointer xlog. We could also fix this by delaying
checkpointer later after gxact->valid resetting in
FinishPreparedTransaction()), but RecordTransactionCommitPrepared() ->
SyncRepWaitForLSN() might be time-consuming or block for some time somehow
(locking, network lag, etc), thus it could block checkpointer for too long time
- that is surely not good. Also it seems that we could fix that by moving
"gxact->valid = false" ahead of delayChkpt resetting, but that is kind of ugly
also, also that is a bit risky for a stable release.

There were two solutions that were discussed previously. One is to use locking
mechanism, but that hurts OLTP performance; Another is to remove the false
positive cases in RecoverPreparedTransactions(), but it is possible that
related clog has been removed by subsequent vacuum operations so it is not
reliable also.

Co-authored-by: Hao Wu <gfphoenix78@gmail.com>

Reviewed-by: Asim R P <pasim@vmware.com>
Reviewed-by: Ashwin Agrawal <aashwin@vmware.com>