Fix the planner crash when the group key is an expression in multi-agg (#10839)
When the group key is an expression, i.e. not a column name, and
there are some other ungrouped targets that reference the group key,
the planner may crash or produce an incorrect result.

Normally, the function check_ungrouped_columns() will scan the
expression trees for ungrouped variables on the target list.
Later, subquery_planner() may transafer the target list to a
new equivalent list, for example,

SELECT dttm >= 0, NOT dttm >= 0 FROM foo GROUP BY 1;

`NOT dttm >= 0` will be converted to `dttm < 0`. In multi-agg,
deconstruct_agg_info() calls deconstruct_expr() to generate
the final target list. Since `dttm < 0` is not an Aggref, it
should reference the group key as the original expression.

The idea to fix this issue is to add the column names to the
output in the first aggregate, so `dttm` could reference the
output of the first aggregate.

Reported-by: github.com/oSkrobuk
Co-authored-by: Jinbao Chen <jinchen@pivotal.io>
Co-authored-by: Weinan Wang <wweinan@vmware.com>