PartitionSelector node is not projection capable
This is how it is on master branch.  Without this patch, it is
possible to generate a plan such as the following:

   ->  Merge Join
         Merge Cond: ((t_left.a = (func(t_right.a))) AND ((t_left.b)::text = (t_right.b)))
         ->  Sort  t_left
         ->  Sort  t_right
               Sort Key: (func(t_right.a)), t_right.b
               ->  Partition Selector for t_left
                     Filter: (func(t_right.a))
                     ->  Broadcast Motion
                           ->  Append
                                 ->  Seq Scan on t_right

The problem with such a plan is the Sort node needs to generate a new
junk attribute representing func(t_rigth.a) by appending an entry to
the target list of its child - Partition Selector.  This change also
applies to Motion node's target list because Motion node shares target
list with Partition Selector.  However, the Append node's target list
does not contain this junk attribute.  This results in the motion sender
sending a tuple with less number of attributes than the receiver's
expectation.

With the proposed patch, prepare_sort_from_pathkeys() creates a new
Result node with its own target list when generating the sort keys.
Any junk attributes added do not propagate to the Result node's
children.

A similar plan was generated at a production deployment, causing
SIGSEGV when the Sort node tried to interpret the tuple returned by
the Partition Selector.  The regress test is a simplified version of the
production plan.

Note that Orca seems to implement Partition Selector nodes differently
and may not suffer from this problem.  The plan generated by Orca for
basic sort and merge query such as

    "select * from part_table order by partition_key;"

contains a Result node between Sort and Partition Selector nodes.

Co-authored-by: Gang Xiong <gangx@vmware.com>
Co-authored-by: Richard Guo <guofenglinux@gmail.com>