Fix aocsseg inconsistency when rollback add column
Before current commit we initailized pg_aocsseg entries with frozen
tuples to make these entries implicitly visible even after rollback.
It is a working approach for AO tables, but AOC tables store additional
"vpinfo" structure with serialized information about every column EOF.
This can cause inconsistency between "vpinfo" structure in pg_aocsseg
entries and the actual number of columns in a table when we modify the
amount of columns under explicit transaction and rollback it later.
As a result we fail on asserts in "GetAllAOCSFileSegInfo_pg_aocsseg_rel()"
or (in a case of a build without asserts) in some unpredictable places
in the code when retrieve metadata from pg_aocsseg.

As a fix this commit inserts simple tuples to pg_aocsseg but still insert
frozen tuples to gp_fastsequence (as it generates row numbers for segment
files and should never be rollbacked to avoid inconsistency in index pointers
of the AO/AOC tables).

Co-authored-by: Vasiliy Ivanov <ivi@arenadata.io>