Fix race condition with unprotected use of a latch pointer variable
This is a follow-up commmit of "Change walreceiver to use its procLatch
instead of a custom latch" to fix a race condition issue introduced by
that. This commit partially backported 45f9d08 from Postgres. Also, the
latch in WalRcvShutdownHandler has been changed to procLatch due to it
can't be protected by spinlock in the signal handler, because it could
cause a deadlock.