gpinitsystem: update default locale on linux
On linux gpinitsystem was defaulting to the hard coded value of
"en_US.utf8". Photon3 uses the value en_US.UTF-8. Later checks in
gpinitsystem would fail due to this discrepancy. Thus, query the system
for the correct utf-8 variation to use as the default.

If no utf8 variant is found on the system default to "en_US.utf8" to
prevent future checks and code within gpinitsystem from failing, and a
friendly error message returned.

Co-authored-by: Jamie McAtamney <jmcatamney@vmware.com>