Make exhaustive2 the default (#11237)
* Change default of optimizer_join_order to exhaustive2

* ICW changes for exhaustive2

* Limit INDF predicate selectivity to 100%

In the case where we joined two histograms that consisted entirely of
NULL values we ended up with a row count estimate that was larger than
the cartesian product of the two tables, which is impossible to achieve
with a join.

This triggered the assert GPOS_ASSERT(CStatistics::MinRows <= scale_factor)
in CJoinStatsProcessor::CalcJoinCardinality().

* MDP changes

Took about 80% of the MDPs that used the old default optimizer_join_order
'exhaustive' and converted them to 'exhaustive2' by altering the traceflags
in the same way a guc change would have. As expected, this lead to some
plan changes, mostly join order changes and different aggregation and
motions. MDPs that were found to have more serious issues are handled in
separate commits.

* Update check for circular dependencies in group expressions

Found a case where we find circular dependencies in a group expression
even after exploring a group, see the MDP that is part of this commit.

* Fix LOJ to inner join conversion issue with DPv2

An unnecessary call to PexprCollapseJoins in
CExpressionPreprocessor::PexprOuterJoinToInnerJoin() prevented us from
converting some LOJs into inner joins by converting them into NAry
joins before PexprOuterJoinToInnerJoin had a chance to turn them into
inner joins.  Note that the preprocessor calls PexprCollapseJoins
right after calling PexprOuterJoinToInnerJoin.

Regenerated 3 MDPs that are meant to test DPv2 with only one join
order enabled (query, mincard or greedy). We now convert the LOJs in
them to inner joins and that wasn't intended. In the distant future,
when we get rid of the DP, query, mincard and greedy xforms we can use
the code path tested in these MDPs to implement different join
orders. Right now the trace flags used in these MDPs are not used
elsewhere.