[Orca] sort includes
I find myself doing it so often that it's time to upstream this: let
clang-format sort and regroup our #include blocks.

We actually tried doing this during the original #10357 but one
logistical hurdle of trying to sort #includes was that our headers are
not exactly order-independent, hence it retreated to bing a TODO
comment.

About 70% of this PR is automated by just running clang-format with the
new configuration, and 30% was manually fixing any order dependency that
the sorting surfaces.

This is spiritually a cherry-pick (actually manually remade) of commit
f81aaafd25206b8c .

(cherry picked from commit f81aaafd25206b8c52205835ba19f6bcb0a66297)