Change the default for optimizer_damping_factor_join to 0 (#11468)
* Change default of optimizer_damping_factor_join to 0.0

* Limit join scale factor to cardinality of larger table

When we treat multiple join predicates as independent, limit
the resulting scale factor to the cardinality of the larger table.
This has the effect of a join condition on a unique key with a
referential integrity constraint. In other words, the estimated
join cardinality is the cardinality of the larger of the two tables.

The general formula for the cardinality of an expression

  R join S on R.c = S.c

is

  |R join S on R.c = S.c| = |R| * |S| / max(NDV(R.c), NDV(S.c))

|x| stands for the cardinality estimate for x.

The expression in the denominator, max(NDV(R.c), NDV(S.c)), is
the scale factor.

For a single equi-join predicate like in this example, the
scale factor can't exceed max(|R|, |S|), since the NDVs can't
be greater than the cardinality of the table.

This commit applies a similar rule when we have multiple scale
factors. For example, look at this join:

  R join S on R.a = S.a and R.b = S.b

If we assume columns a and b are not correlated, then we'll just
multiply the selectivities:

  |R| * |S| / (max(NDV(R.a), NDV(S.a)) * max(NDV(R.b), NDV(S.b)))

In this formula, we want to make sure that the scale factor
max(NDV(R.a), NDV(S.a)) * max(NDV(R.b), NDV(S.b)))
doesn't exceed max (|R|, |S|).

In other words, we assume that the join predicate

  R.a = S.a and R.b = S.b

selects 1 or more rows from the larger table, given a row from the
smaller table. We make the same assumption when we only have a
single join predicate like in the first example, that's the
assumption the formula for the cardinality is based on.

* MDP changes

Changed the setup of some MDPs to be more realistic, with RI
constraints between the tables. Updated the expectations in
some cases. Minor cardinality changes in some other cases.

* Convert an arbitrary 40% of MDPs to use damping factor 0

Of all the MDPs that currently use a join damping factor of 0.01, pick
2 of every 5 (a total of 373) and change the damping factor from 0.01
to 0.

Note that this is in addition to the MDPs that already test join
damping factor 0 explicitly.