Limit anyarray coercion to INSERT statements
The anyarray coercion hack was added circa 2008 to accommodate INSERT
statements into catalog tables such as pg_statistic, which sport the
anyarray type - a pseudo-type that does not handle insertions.

However, coerce_type() is a very low level function that has a variety
of callsites. The logic added has an adverse effect on views, for e.g.
that have an OpExpr with array types:

postgres=# CREATE VIEW fooview AS SELECT '{1}'::int[] = '{2}'::int[];

postgres=# \d+ fooview
                 View "public.fooview"
  Column  |  Type   | Modifiers | Storage | Description
----------+---------+-----------+---------+-------------
 ?column? | boolean |           | plain   |
View definition:
 SELECT '{1}'::anyarray = '{2}'::anyarray;

We can see that the view definition includes incorrect downcasts to the
pseudo-type anyarray. (It should have been
SELECT '{1}'::integer = '{2}'::integer;)

The type information has been lost forever and thus running the view
definition inevitably yields:

postgres=# SELECT '{1}'::anyarray = '{2}'::anyarray;
ERROR:  cannot accept a value of type anyarray
LINE 1: SELECT '{1}'::anyarray = '{2}'::anyarray;

We fix this issue by restricting the scope of the type coercion logic to
INSERTs only, in line with what was originally intended.

This is not the ideal fix - ideally, we should rip out the hack from
coerce_type() completely and adapt our utilities to emit array_in
expressions to insert/update columns with anyarray types (as
demonstrated in the changes to bfv_statistic.sql). This is what we will
be doing in future versions of GPDB.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>