CI: correct jinja2 if/endif in gpdb-tpl.yml from 0668db73263
The commit 0668db73263 introduced if/endif jinja2 blocks in the CI
template file.  Those blocks incorrectly associated most of the file
with the added "sles12" options.  This caused gen_pipeline.py to fail
for targets other than "prod".

This commit fixes that.

There are no changes in the 6X prod pipeline generated files(other that
the generated timestamp), so those are omitted from this comment.

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>
Co-authored-by: Brent Doil <bdoil@vmware.com>