Add a new dblink function 'dblink_connect_no_auth' (#11566)
* Add a new dblink function 'dblink_connect_no_auth'

dblink_connect_no_auth() is identical to dblink_connect_u(),
except that it bypass authentication when gpdb uses it to connect
to itself, so impersonation and subsequent escalation of privileges
can occur when using it.

How does it work?

We make dblink_connect_no_auth linked to backend_fe-connect.o,
and undef the FRONTEND in dblink_no_auth.c. So dblink can use
internal gpdb connection to connect to local master and then the
authentication will be skipped. More details are in the function
ClientAuthentication.

Only dblink_connect_no_auth() uses internal gpdb connection, other
dblink functions still use the frontend libpq connection.
This function is really hacked and has some serious security issues, why do we add it?
Because before the commit 1dc23be, dblink is wrongly linked to backend libpq,
and when using it to connect to master itself,
authentication is skipped. However, after the link issue fixed,
customers still want this feature, so we have to add it back.

Co-authored-by: Chen Mulong chenmulong@gmail.com