Moving open partition bounds check into individual comparisons (#11602)
* Removing or renaming flags for open bounds checking
* Add a check for open partition bounds in PdxlnCmp

Method CTranslatorExprToDXLUtils::PdxlnCmp() is a convenient
place to add the check for open boundaries close to the actual
comparison, so we don't have to do this check at the top level
of the predicate.

* Change the partition elimination expr for IS [NOT] NULL

Changing this expression for range-partitioned tables. I didn't
understand the existing logic. As far as I can tell, the NULL value
can only go into the default range partition, but the default
partition can also have non-null values (we probably don't want to
write code that detects when this is not possible).

Therefore, an IS NULL predicate selects the default range partition
while an IS NOT NULL predicate selects all range partitions.

* Plan changes, mainly partition filter exprs
* Add MDP for IS NULL predicate