Fix use after free in GetSnapshotData()
After a call to repalloc the address of the newly allocated memory must
be saved back in the pointer, otherwise the pointer is left to point to
the old (now free’d) memory.
In GetSnapshotData(), GPDB repalloc()'s comboCids and then memcpy()'s
into it contents larger than than its previously allocated size.
This overflows to other allocations in the block and corrupts the chunk
header of other allocations.
In the attached test, this corrupt the chunk header of ActiveSnapshot
and subsequently crashes the backend later during PopSnapshot().

The affected allocation happens to be that of the ActiveSnapshot, but
that is coincidental. This could affect any allocation in the same block
of the TopTransactionContext.

To reproduce this, we first need to execute the query within a
transaction which activates the whole comboCids mechanism. Then, we need
a plan with at least two slices that execute on the segment with the
following sequence of events in this order:
1. Writer slice takes a snapshot & updates shared memory.
2. Reader slice takes a snapshot & updates its usedComboCids to a
   non-zero value
3. Writer slice takes an additional snapshot and updates the shared
   memory.
4. Reader slice takes another snapshot & updates usedComboCids. This
   will hit the faulty case.

The repro in the attached test achieves with a CREATE TABLE AS with:
1. A toast-able column (e.g text, which creates a toast table that
   increments command counter before execution of the SELECT)
2. Calls to SQL UDFs (which in turn call GetSnapshotData())
3. Generating a plan with 2 slices that will execute this SQL UDF on
   both the writer & reader slice (resulting in the sequence of events
   described above).

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>