Move all of gp_sparse_vector staff to sparse_vector schema
Reason: the gp_sparse_vector has many functions and operators
	are the same as GPDB/PG built-in functions and operators.

Currently, the extension will create its functions and
operators in the public schema and will override the built-in
functions and operators. As the behavior between extension
functions and operators are different with built-in
functions and operators. Hence, it will cause some issues.

In this commit, the gp_sparse_vector's functions and operators will be
put into a separated schema to split extension and built-in.
All tests are also updated.