Fix incorrect group pathkey
For query like 'select 1, a from t group by grouping sets(1,2)', we assume that
there is a one-to-one mapping between the pathkeys in root->group_pathkeys and
the group columns, so we add a workaround to construnct a one-to-one pathkey
list even it contains a constant pathkey. In this case, function
'prepare_sort_from_pathkeys' will consider the pathkey list has no sort column.
To fix this, we remove the constant pathkey and then pass it to
'prepare_sort_from_pathkeys'. We should resolve the GPDB_91_MERGE_FIXME in
'plan_grouping_extension' to come up with a thorough fix.
This issue only exist on 6X.