Fix incremental recovery failure because the checkpoint redo wal file before divergence LSN is gone.
gprecoverseg calls pg_rewind to do incremental recovery. For unclean shutdown,
pg_rewind calls single mode postgres to ensure the clean shutdown, however the
single mode postgres would introduce two additional checkpoints, one is after
crash recovery, and another is for shutdown. Current gp6 xlog removing/recyling
logic ensures that the previous checkpoint.redo is kept when doing checkpoint, now
that two additional checkpoints are added the checkpoint wal file before the
divergence LSN might be recycled/removed and thus pg_rewind fails sometimes due
to missing wal files. This issue happens also when the primary is not actually
running but just was marked down by fts. The coordinator promotes the mirror,
and the continuing checkpoint on the primary could remove the needed wal files
for pg_rewind.

In real scenario this could be reproduced by the below script,

  psql -c 'create table t(a int);'
  psql -c 'insert into t select * from generate_series(1,10000000);'
  psql -c 'insert into t select * from generate_series(1,10000000);'
  psql -c 'insert into t select * from generate_series(1,10000000);'

  kill -9 `ps -ef |grep 6002|grep gpAux|awk '{print $2}'`
  psql -c "select gp_request_fts_probe_scan()" postgres

  gprecoverseg -av
  echo $?

For this case, the key point to reproduce this issue is there is an ongoing
checkpoint after bulk inserting, then if the primary postmaster is kill-9-ed,
the checkpointer flushes the replication slots in CheckPointReplicationSlots()
already but waits in CheckPointBuffers() for some time since bgwriter is gone.
The checkpoint might finish or not (checkpoint xlog write and xlog file
recycling/removal) before gprecoverseg, but either scenario could lead to the
issue (since this checkpoint xlog is after divergence LSN so it is uselss -
pg_rewind needs to find the common ancestor).

This issue happens on master and upstream also. Fixing this in 6X at first by
trying to keep the checkpoint.redo before the oldest replication slot LSN.

Co-authored-by: Gang Xiong <gxiong@pivotal.io>

Reviewed-by: Asim R P <pasim@vmware.com>
Reviewed by: Ashwin Agrawal <aashwin@vmware.com>