Make gp_toolkit bloat computation more lenient (#11787)
The gp_toolkit.gp_bloat_diag view uses average width of a row as
computed from a sample of a relation's blocks (by analyze) and
multiplies it with reltuples to compute expected size of the relation.
The width itself is approximate and using it as a multiplying factor
further increases the error.  Consequently, we have seen confusion in
production deployments where the gp_bloat_diag reported "moderate
bloat" for a relation right after it was vacuumed using "vacuum full".

To fix this confusion, a factor is added to the approximate width
computed from pg_statistic.  The added amount is proportional to the
number of variable length attributes in the relation.  The thresholds
used to report moderate and severe bloat are changed too, so that the
bloat compulation now errors on the side of "not vacuuming".

Co-authored-by: Jinbao Chen <cjinbao@vmware.com>
Reviewed by: Ashwin Agrawal <aashwin@vmware.com>

Co-authored-by: Asim R P <pasim@vmware.com>