gpconfig: aborting a config update with down host should exit properly (#11765)
When a host is down, and a user runs gpconfig to update any config: we
warn them and ask for confirmation to continue with the config update.
If the user chooses to abort the update, we currently do not exit out
properly due to (non daemon) threads staying open when exiting.

This change fixes the abort/exit to close threads

We also remove use of GpHostCache as part of this change and instead use
functions from GpArray

* behave: refactor gpconfig behave tests steps

We need to verify whether we did not change the config file.  This can
occur when a host is unreachable and the user chooses to abort the
config update.