Fix exec location INITPLAN not showing up in \df+
When running \df+ on a function that has exec location INITPLAN, the
"Execute on" column did not properly display INITPLAN. This is because
psql was not updated to recognize the new exec location. To fix, simply
add exec location INITPLAN to the \df+ query.

commit reference:
https://github.com/greenplum-db/gpdb/commit/a21ff23b615e5f9a7dd3b02d82fad86175b188ec

Co-authored-by: Kevin Yeap <kyeap@vmware.com>
Co-authored-by: Jimmy Yih <jyih@vmware.com>

Backport of master commit:
https://github.com/greenplum-db/gpdb/commit/e0a93797eae23396a1967002208e622139c5f849